Refactoring
create specific files and organize code better
